rename: 쿠폰 지급 관련 컴포넌트 CouponProvision 폴더로 이동
