+++ src/pages/CouponProvisionPage.tsx
@@ -1,5 +1,5 @@
-import CouponProvisionHeader from "@/components/common/Header/CouponProvisionHeader";
-import CouponProvisionInfoTable from "@/components/common/Table/CouponProvisionInfoTable";
+import CouponProvisionHeader from "@/components/CouponProvision/CouponProvisionHeader";
+import CouponProvisionInfoTable from "@/components/CouponProvision/CouponProvisionInfoTable";
 import Title from "@/components/common/Title";
 import CouponProvisionContextProvider from "@/contexts/CouponProvisionContext";
